requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
Django==5.2.2
django-oidc-provider==0.8.4
djangorestframework==3.16.0
drf-spectacular==0.28.0
future==1.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
pillow==11.2.1
pip==24.2
psycopg2-binary==2.9.10
pycryptodomex==3.23.0
pyjwkest==1.4.2
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.4
rpds-py==0.25.1
djangorestframework-simplejwt==5.3.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.9.0